move weird keys to escape package for re-use
